requirements.txt
flask
flask-cors
flask-bcrypt
flask-mail
mysql-connector-python
python-dotenv
requests
Pillow